# backend/src/generated/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
			"TaskStatus": {
				"enum": [
					"pending",
					"in_progress",
					"completed",
					"cancelled"
				],
				"type": "string"
			},
			"TaskPriority": {
				"enum": [
					"low",
					"medium",
					"high",
					"urgent"
				],
				"type": "string"
			},
			"Record_string.unknown_": {
				"properties": {},
				"additionalProperties": {},
				"type": "object",
				"description": "Construct a type with a set of properties K of type T"
			},
			"User": {
				"properties": {
					"id": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"password": {
						"type": "string"
					},
					"role": {
						"type": "string"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"id",
					"name",
					"email",
					"password",
					"role",
					"createdAt",
					"updatedAt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Task": {
				"properties": {
					"id": {
						"type": "string"
					},
					"title": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"status": {
						"$ref": "#/components/schemas/TaskStatus"
					},
					"priority": {
						"$ref": "#/components/schemas/TaskPriority"
					},
					"dueDate": {
						"type": "string",
						"format": "date-time"
					},
					"completedAt": {
						"type": "string",
						"format": "date-time"
					},
					"recurrenceRule": {
						"type": "string"
					},
					"tags": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"metadata": {
						"$ref": "#/components/schemas/Record_string.unknown_"
					},
					"userId": {
						"type": "string"
					},
					"user": {
						"$ref": "#/components/schemas/User"
					},
					"projectId": {
						"type": "string"
					},
					"project": {
						"$ref": "#/components/schemas/Project"
					},
					"assignedToId": {
						"type": "string"
					},
					"assignedTo": {
						"$ref": "#/components/schemas/User"
					},
					"order": {
						"type": "number",
						"format": "double"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"id",
					"title",
					"status",
					"priority",
					"userId",
					"user",
					"order",
					"createdAt",
					"updatedAt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Project": {
				"properties": {
					"id": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"color": {
						"type": "string"
					},
					"icon": {
						"type": "string"
					},
					"isShared": {
						"type": "boolean"
					},
					"isArchived": {
						"type": "boolean"
					},
					"metadata": {
						"$ref": "#/components/schemas/Record_string.unknown_"
					},
					"ownerId": {
						"type": "string"
					},
					"owner": {
						"$ref": "#/components/schemas/User"
					},
					"tasks": {
						"items": {
							"$ref": "#/components/schemas/Task"
						},
						"type": "array"
					},
					"members": {
						"items": {
							"$ref": "#/components/schemas/User"
						},
						"type": "array"
					},
					"order": {
						"type": "number",
						"format": "double"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"id",
					"name",
					"color",
					"isShared",
					"isArchived",
					"ownerId",
					"owner",
					"tasks",
					"order",
					"createdAt",
					"updatedAt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ApiResponse_Task_": {
				"description": "Standard API Response wrapper\nUsed by all TSOA controllers",
				"properties": {
					"success": {
						"type": "boolean"
					},
					"message": {
						"type": "string"
					},
					"data": {
						"$ref": "#/components/schemas/Task"
					},
					"timestamp": {
						"type": "string"
					}
				},
				"required": [
					"success",
					"message",
					"timestamp"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CreateTaskDto": {
				"properties": {
					"title": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"status": {
						"$ref": "#/components/schemas/TaskStatus"
					},
					"priority": {
						"$ref": "#/components/schemas/TaskPriority"
					},
					"dueDate": {
						"anyOf": [
							{
								"type": "string",
								"format": "date-time"
							},
							{
								"type": "string"
							}
						]
					},
					"projectId": {
						"type": "string"
					},
					"assignedToId": {
						"type": "string"
					},
					"recurrenceRule": {
						"type": "string"
					},
					"tags": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"metadata": {
						"$ref": "#/components/schemas/Record_string.unknown_"
					}
				},
				"required": [
					"title"
				],
				"type": "object",
				"additionalProperties": false
			},
			"PaginatedResponse_Task-Array_": {
				"description": "Paginated API Response wrapper\nUsed for endpoints that return lists with pagination",
				"properties": {
					"success": {
						"type": "boolean"
					},
					"message": {
						"type": "string"
					},
					"data": {
						"items": {
							"$ref": "#/components/schemas/Task"
						},
						"type": "array"
					},
					"pagination": {
						"properties": {
							"totalPages": {
								"type": "number",
								"format": "double"
							},
							"total": {
								"type": "number",
								"format": "double"
							},
							"limit": {
								"type": "number",
								"format": "double"
							},
							"page": {
								"type": "number",
								"format": "double"
							}
						},
						"required": [
							"totalPages",
							"total",
							"limit",
							"page"
						],
						"type": "object"
					},
					"timestamp": {
						"type": "string"
					}
				},
				"required": [
					"success",
					"message",
					"data",
					"pagination",
					"timestamp"
				],
				"type": "object",
				"additionalProperties": false
			},
			"TaskStatsDto": {
				"properties": {
					"total": {
						"type": "number",
						"format": "double"
					},
					"pending": {
						"type": "number",
						"format": "double"
					},
					"inProgress": {
						"type": "number",
						"format": "double"
					},
					"completed": {
						"type": "number",
						"format": "double"
					},
					"cancelled": {
						"type": "number",
						"format": "double"
					},
					"overdue": {
						"type": "number",
						"format": "double"
					},
					"dueToday": {
						"type": "number",
						"format": "double"
					},
					"dueSoon": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"total",
					"pending",
					"inProgress",
					"completed",
					"cancelled",
					"overdue",
					"dueToday",
					"dueSoon"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ApiResponse_TaskStatsDto_": {
				"description": "Standard API Response wrapper\nUsed by all TSOA controllers",
				"properties": {
					"success": {
						"type": "boolean"
					},
					"message": {
						"type": "string"
					},
					"data": {
						"$ref": "#/components/schemas/TaskStatsDto"
					},
					"timestamp": {
						"type": "string"
					}
				},
				"required": [
					"success",
					"message",
					"timestamp"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UpdateTaskDto": {
				"properties": {
					"title": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"status": {
						"$ref": "#/components/schemas/TaskStatus"
					},
					"priority": {
						"$ref": "#/components/schemas/TaskPriority"
					},
					"dueDate": {
						"anyOf": [
							{
								"type": "string",
								"format": "date-time"
							},
							{
								"type": "string"
							}
						],
						"nullable": true
					},
					"completedAt": {
						"anyOf": [
							{
								"type": "string",
								"format": "date-time"
							},
							{
								"type": "string"
							}
						],
						"nullable": true
					},
					"projectId": {
						"type": "string",
						"nullable": true
					},
					"assignedToId": {
						"type": "string",
						"nullable": true
					},
					"recurrenceRule": {
						"type": "string",
						"nullable": true
					},
					"tags": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"metadata": {
						"$ref": "#/components/schemas/Record_string.unknown_"
					},
					"order": {
						"type": "number",
						"format": "double"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"ApiResponse_null_": {
				"description": "Standard API Response wrapper\nUsed by all TSOA controllers",
				"properties": {
					"success": {
						"type": "boolean"
					},
					"message": {
						"type": "string"
					},
					"data": {
						"type": "number",
						"enum": [
							null
						],
						"nullable": true
					},
					"timestamp": {
						"type": "string"
					}
				},
				"required": [
					"success",
					"message",
					"timestamp"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ApiResponse_Task-Array_": {
				"description": "Standard API Response wrapper\nUsed by all TSOA controllers",
				"properties": {
					"success": {
						"type": "boolean"
					},
					"message": {
						"type": "string"
					},
					"data": {
						"items": {
							"$ref": "#/components/schemas/Task"
						},
						"type": "array"
					},
					"timestamp": {
						"type": "string"
					}
				},
				"required": [
					"success",
					"message",
					"timestamp"
				],
				"type": "object",
				"additionalProperties": false
			},
			"BulkUpdateTasksDto": {
				"properties": {
					"taskIds": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"updates": {
						"properties": {
							"tags": {
								"items": {
									"type": "string"
								},
								"type": "array"
							},
							"assignedToId": {
								"type": "string",
								"nullable": true
							},
							"projectId": {
								"type": "string",
								"nullable": true
							},
							"priority": {
								"$ref": "#/components/schemas/TaskPriority"
							},
							"status": {
								"$ref": "#/components/schemas/TaskStatus"
							}
						},
						"type": "object"
					}
				},
				"required": [
					"taskIds",
					"updates"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ReorderTasksDto": {
				"properties": {
					"taskOrders": {
						"items": {
							"properties": {
								"order": {
									"type": "number",
									"format": "double"
								},
								"id": {
									"type": "string"
								}
							},
							"required": [
								"order",
								"id"
							],
							"type": "object"
						},
						"type": "array"
					}
				},
				"required": [
					"taskOrders"
				],
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {
			"jwt": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "JWT"
			}
		}
	},
	"info": {
		"title": "Todo AI API",
		"version": "1.0.0",
		"description": "Auto-generated OpenAPI specification for Todo AI",
		"license": {
			"name": "MIT"
		},
		"contact": {
			"name": "API Support",
			"email": "support@todoai.com"
		}
	},
	"paths": {
		"/api/tasks": {
			"post": {
				"operationId": "CreateTask",
				"responses": {
					"201": {
						"description": "Task created successfully",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_Task_"
								}
							}
						}
					}
				},
				"description": "Create a new task",
				"summary": "Create a task",
				"tags": [
					"Tasks"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateTaskDto"
							},
							"example": {
								"title": "Complete project documentation",
								"description": "Write comprehensive docs",
								"status": "pending",
								"priority": "high",
								"dueDate": "2025-10-25T10:00:00Z",
								"projectId": "123e4567-e89b-12d3-a456-426614174000",
								"tags": [
									"documentation",
									"priority"
								]
							}
						}
					}
				}
			},
			"get": {
				"operationId": "GetTasks",
				"responses": {
					"200": {
						"description": "Tasks retrieved successfully",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/PaginatedResponse_Task-Array_"
								}
							}
						}
					}
				},
				"description": "Get all tasks with filtering and pagination",
				"summary": "List all tasks",
				"tags": [
					"Tasks"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "page",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "search",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "status",
						"required": false,
						"schema": {
							"$ref": "#/components/schemas/TaskStatus"
						}
					},
					{
						"in": "query",
						"name": "priority",
						"required": false,
						"schema": {
							"$ref": "#/components/schemas/TaskPriority"
						}
					},
					{
						"in": "query",
						"name": "projectId",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "assignedToId",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "dueBefore",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "dueAfter",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "tags",
						"required": false,
						"schema": {
							"type": "array",
							"items": {
								"type": "string"
							}
						}
					},
					{
						"in": "query",
						"name": "sortBy",
						"required": false,
						"schema": {
							"type": "string",
							"enum": [
								"title",
								"createdAt",
								"updatedAt",
								"dueDate",
								"priority",
								"order"
							]
						}
					},
					{
						"in": "query",
						"name": "sortOrder",
						"required": false,
						"schema": {
							"type": "string",
							"enum": [
								"ASC",
								"DESC"
							]
						}
					}
				]
			}
		},
		"/api/tasks/stats": {
			"get": {
				"operationId": "GetTaskStats",
				"responses": {
					"200": {
						"description": "Task statistics retrieved successfully",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_TaskStatsDto_"
								}
							}
						}
					}
				},
				"description": "Get task statistics",
				"summary": "Get task statistics",
				"tags": [
					"Tasks"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": []
			}
		},
		"/api/tasks/{id}": {
			"get": {
				"operationId": "GetTaskById",
				"responses": {
					"200": {
						"description": "Task retrieved successfully",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_Task_"
								}
							}
						}
					}
				},
				"description": "Get a single task by ID",
				"summary": "Get task by ID",
				"tags": [
					"Tasks"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"description": "Task ID",
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"put": {
				"operationId": "UpdateTask",
				"responses": {
					"200": {
						"description": "Task updated successfully",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_Task_"
								}
							}
						}
					}
				},
				"description": "Update a task",
				"summary": "Update task",
				"tags": [
					"Tasks"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"description": "Task ID",
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateTaskDto"
							},
							"example": {
								"title": "Updated task title",
								"status": "in_progress",
								"priority": "urgent"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteTask",
				"responses": {
					"200": {
						"description": "Task deleted successfully",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_null_"
								}
							}
						}
					}
				},
				"description": "Delete a task",
				"summary": "Delete task",
				"tags": [
					"Tasks"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"description": "Task ID",
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/tasks/bulk": {
			"patch": {
				"operationId": "BulkUpdateTasks",
				"responses": {
					"200": {
						"description": "Tasks updated successfully",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_Task-Array_"
								}
							}
						}
					}
				},
				"description": "Bulk update tasks",
				"summary": "Bulk update multiple tasks",
				"tags": [
					"Tasks"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/BulkUpdateTasksDto"
							},
							"example": {
								"taskIds": [
									"task-id-1",
									"task-id-2"
								],
								"updates": {
									"status": "completed",
									"priority": "low"
								}
							}
						}
					}
				}
			}
		},
		"/api/tasks/reorder": {
			"patch": {
				"operationId": "ReorderTasks",
				"responses": {
					"200": {
						"description": "Tasks reordered successfully",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_null_"
								}
							}
						}
					}
				},
				"description": "Reorder tasks",
				"summary": "Reorder tasks",
				"tags": [
					"Tasks"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ReorderTasksDto"
							},
							"example": {
								"taskOrders": [
									{
										"id": "task-id-1",
										"order": 0
									},
									{
										"id": "task-id-2",
										"order": 1
									}
								]
							}
						}
					}
				}
			}
		},
		"/projects/{projectId}/tasks": {
			"get": {
				"operationId": "GetTasksByProject",
				"responses": {
					"200": {
						"description": "Tasks retrieved successfully",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/PaginatedResponse_Task-Array_"
								}
							}
						}
					}
				},
				"description": "Get tasks by project",
				"summary": "Get all tasks in a project",
				"tags": [
					"Tasks"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"description": "Project ID",
						"in": "path",
						"name": "projectId",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "page",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "search",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "status",
						"required": false,
						"schema": {
							"$ref": "#/components/schemas/TaskStatus"
						}
					},
					{
						"in": "query",
						"name": "priority",
						"required": false,
						"schema": {
							"$ref": "#/components/schemas/TaskPriority"
						}
					},
					{
						"in": "query",
						"name": "assignedToId",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "dueBefore",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "dueAfter",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "tags",
						"required": false,
						"schema": {
							"type": "array",
							"items": {
								"type": "string"
							}
						}
					},
					{
						"in": "query",
						"name": "sortBy",
						"required": false,
						"schema": {
							"type": "string",
							"enum": [
								"title",
								"createdAt",
								"updatedAt",
								"dueDate",
								"priority",
								"order"
							]
						}
					},
					{
						"in": "query",
						"name": "sortOrder",
						"required": false,
						"schema": {
							"type": "string",
							"enum": [
								"ASC",
								"DESC"
							]
						}
					}
				]
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	],
...
}